requirements.txt
# requirements.txt
# Add or remove packages as needed for your project.

Pillow>=9.0
numpy>=1.24
tqdm>=4.0
